CleanUp const.py + Factory
