+++ pysolarfocus/component_factory.py
@@ -1,0 +1,33 @@
+from .components.heating_circuit import *
+from .components.boiler import *
+from .components.heat_pump import *
+from .components.buffer import *
+from .components.pellets_boiler import *
+from .components.photovoltaic import *
+from .const import Systems
+
+
+class ComponentFactory:
+    @staticmethod
+    def heating_circuit(system:Systems)->HeatingCircuit:
+        return HeatingCircuit()
+    
+    @staticmethod
+    def boiler(system:Systems)->Boiler:
+        return Boiler()
+    
+    @staticmethod
+    def heatpump(system:Systems)->HeatPump:
+        return HeatPump()
+    
+    @staticmethod
+    def photovoltaic(system:Systems)->Photovoltaic:
+        return Photovoltaic()
+    
+    @staticmethod
+    def pelletsboiler(system:Systems)->PelletsBoiler:
+        return PelletsBoiler()
+    
+    @staticmethod
+    def buffer(system:Systems)->Buffer:
+        return Buffer()
